+++ src/allocator/buffer.rs
@@ -1,13 +1,5 @@
-#![allow(dead_code)]
-
 use std::alloc::{self, Layout, LayoutError};
 use std::ptr::NonNull;
-
-use tokio::sync::mpsc;
-
-// TODO(i.erin) add miri tests
-// TODO(i.erin) more tests!!!
-// TODO(i.erin) iterator over buffer chain?
 
 /// Instrusive linked list heap allocated node layout.
 ///
@@ -23,7 +15,7 @@ struct BufferLayout {
 }
 
 /// Pointer to the heap allocated intrusive linked list node.
-struct Buffer(NonNull<BufferLayout>);
+pub struct Buffer(NonNull<BufferLayout>);
 
 impl Buffer {
     /// Allocates intrusive linked list node ([`BufferLayout`]) and return pointer to it.
@@ -99,7 +91,7 @@ impl Buffer {
     }
 
     /// Return raw pointer to heap allocated linked list node.
-    pub fn as_ptr(&self) -> *const BufferLayout {
+    fn as_ptr(&self) -> *const BufferLayout {
         self.0.as_ptr()
     }
 
@@ -118,106 +110,6 @@ impl Buffer {
         let (full_layout, _) = layout.extend(payload_layout)?;
 
         Ok(full_layout.pad_to_align())
-    }
-}
-
-type Sender = mpsc::Sender<Buffer>;
-type Receiver = mpsc::Receiver<Buffer>;
-
-struct Chain {
-    buffer: Option<Buffer>,
-    sender: Sender,
-}
-
-impl Chain {
-    /// Crates new instrusive linked chain of buffers.
-    pub fn new(sender: Sender) -> Self {
-        Self { buffer: None, sender }
-    }
-
-    /// Push new intrusive linked list node at head.
-    pub fn push_head(&mut self, mut buffer: Buffer) {
-        assert!((*buffer.mut_next()).is_none());
-
-        *buffer.mut_next() = self.buffer.take();
-        self.buffer = Some(buffer)
-    }
-
-    /// Represent linked list chain as vector of mutable slices.
-    pub fn to_vec(Self { buffer, .. }: &mut Self) -> Vec<&mut [u8]> {
-        let mut result = Vec::new();
-
-        let mut current = buffer;
-        loop {
-            match current {
-                Some(buffer) => {
-                    let (as_mut, next) = buffer.as_mut_and_next();
-                    current = next;
-
-                    result.push(as_mut)
-                }
-                None => return result,
-            }
-        }
-    }
-
-    /// Deallocate in blocking manner.
-    ///
-    /// May deadlock in async context.
-    pub fn blocking_dealloc(Self { buffer: mut next_buffer, sender }: Self) {
-        while let Some(mut buffer) = next_buffer {
-            next_buffer = buffer.mut_next().take();
-            sender.blocking_send(buffer).expect("can't send blocking");
-        }
-    }
-
-    /// Deallocate in async manner.
-    pub async fn dealloc(Self { buffer: mut next_buffer, sender }: Self) {
-        while let Some(mut buffer) = next_buffer {
-            next_buffer = buffer.mut_next().take();
-            sender.send(buffer).await.expect("can't send async");
-        }
-    }
-}
-
-/// Allocates instrusive linked [`Chain`].
-struct Allocator {
-    receiver: Receiver,
-    sender: Sender,
-    capacity: usize,
-}
-
-impl Allocator {
-    /// Crates new allocator with specified buffer size and count.
-    ///
-    /// # Parameters
-    ///
-    /// * `size` --- size of each buffer
-    /// * `count` --- counts of buffer
-    pub async fn new(size: usize, count: usize) -> Self {
-        let (sender, receiver) = mpsc::channel::<Buffer>(count);
-
-        for _ in 0..count {
-            let buffer = Buffer::alloc(size);
-            sender.send(buffer).await.expect("cannot init buffers");
-        }
-
-        Self { sender, receiver, capacity: size * count }
-    }
-
-    /// Allocates [`Chain`] at least the specified size.
-    pub async fn alloc(&mut self, mut size: usize) -> Chain {
-        assert!(size < self.capacity);
-
-        let mut chain = Chain::new(self.sender.clone());
-        while size > 0 {
-            let buffer = self.receiver.recv().await.expect("channel not to be closed");
-
-            size = size.saturating_sub(buffer.len());
-            chain.push_head(buffer);
-        }
-
-        chain
     }
 }
 
@@ -247,22 +139,5 @@ mod test {
         unsafe {
             buffer.dealloc();
         }
-    }
-
-    // Checks that:
-    // TODO(more test!!!)
-    #[test]
-    fn buffer_chain() {
-        const BUFFER_COUNT: usize = 4;
-        const SIZE: usize = 12345;
-
-        let (sender, _reciever) = mpsc::channel(BUFFER_COUNT);
-        let mut buffer_chain = Chain::new(sender);
-
-        buffer_chain.push_head(Buffer::alloc(SIZE));
-        buffer_chain.push_head(Buffer::alloc(SIZE));
-
-        let vec = Chain::to_vec(&mut buffer_chain);
-        assert_eq!(vec.len(), 2);
     }
 }
